requirements.txt
# Core testing framework
pytest>=7.4.0
pytest-playwright>=0.4.0
playwright>=1.37.0

# Data validation
pydantic>=2.0.0
PyYAML>=6.0.0

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pytest-cov>=4.1.0
